Upgrade krama-api to Laravel 12 (from 8.83)
Dependency matrix -> L12: laravel/framework ^12, sanctum ^4, tinker ^2.10,
jwt-auth ^2.8 (2.8.3 supports L12 + PHP 8.2), guzzle ^7.8; removed
fruitcake/laravel-cors (built-in HandleCors now) and facade/ignition; dev:
larastan/larastan ^3, collision ^8, phpunit ^11. platform pinned php 8.2.32
(matches server); minimum-stability stable.

Code fixes:
- Http/Kernel.php: Fruitcake\Cors\HandleCors -> Illuminate\Http\Middleware\HandleCors.
- Carbon 3 strictness (rejects string args): (int) cast jwt ttl/refresh_ttl in
  config/jwt.php + addMinutes/addDays call sites in Auth/Job/Payment controllers.
  This was the login 500.

Verified on PHP 8.3 (serve :8001, shared dev DB): login/me/refresh 200,
employer authz 200, public endpoints 200, scheduler lists all 7 crons,
202 routes, migrate:status clean. Kept classic Kernel skeleton (still honored).

+++ krama-api/app/Http/Controllers/AuthController.php
@@ -587,7 +587,7 @@ class AuthController extends Controller
             'user_id'    => $user->id,
             'type'       => 'refresh',
             'token_hash' => $hash,
-            'expires_at' => now()->addMinutes(config('jwt.refresh_ttl')),
+            'expires_at' => now()->addMinutes((int) config('jwt.refresh_ttl')),
             'created_at' => now(),
         ]);
